+++ CHANGELOG.md
@@ -70,6 +70,8 @@
 - Enhanced `Export-GlobalWsaJsonSpec` cmdlet to include the export from Planning and Preparations Workbook the Workspace ONE Access database IP.
 - Enhanced `Register-vROPSManagementPack` cmdlet to handle the new VCFDiagnostics management pack in VMware Aria Operations.
 - Enhanced `Invoke-GlobalWsaDeployment` cmdlet to ensure that the Workspace ONE Access identity provider is referencing the cluster FQDN.
+- Enhanced `Invoke-UndoIomDeployment` cmdlet to remove the load balancer configured in VMware Aria Suite Lifecycle for VMware Aria Operations.
+- Enhanced `Invoke-UndoPcaDeployment` cmdlet to remove the load balancer configured in VMware Aria Suite Lifecycle for VMware Aria Automation.
 - Converted `aria-operations-notifications-vcf.csv` to 'aria-operations-notifications-vcf.json'.
 - Converted `aria-operations-notifications-srm.csv' to 'aria-operations-notifications-srm.json'.